Changed all size_t to int and 3 unit support

+++ src/bs/bs.h
@@ -12,7 +12,7 @@ namespace bs
 {
     std::vector<double> seq_length(double const begin, double const end, int const size);
     double              span(std::vector<double> const &first, std::vector<double> const &second);
-    size_t              min_sum(std::vector<double> const &data, double const minSum);
+    int                 min_sum(std::vector<double> const &data, double const minSum);
     double              approximate_gain(std::vector<double> const &vec1, std::vector<double> const &vec2);
 }
