+++ src/init/settransitiongamma.cpp
@@ -24,7 +24,7 @@ void Init::setTransitionGamma()
   d_prop2 = vector<vector<num_t>> (d_nRates, vector<num_t> (d_nStates, 0));
   
   // Fill the transion vectors for each action (= production rate)
-  for (size_t adx = 0; adx != d_nRates; ++adx)
+  for (int adx = 0; adx != d_nRates; ++adx)
   {
     num_t const shape = d_dT * (d_mu * d_mu) / d_var;
     num_t const scale = d_detRates[adx] * (d_var / d_mu);
@@ -38,12 +38,12 @@ void Init::setTransitionGamma()
       boost::math::gamma_distribution<num_t> distribution(shape, scale);
       d_prop1[adx][0] = cdf(distribution, 0.5 * d_dL);
       
-      for (size_t idx = 1; idx != d_nStates; ++idx)
+      for (int idx = 1; idx != d_nStates; ++idx)
         d_prop1[adx][idx] = cdf(distribution, (idx + 0.5) * d_dL) - cdf(distribution, (idx - 0.5) * d_dL);
     }
     
     // Probability of failure
-    for (size_t idx = 0; idx != d_nStates; ++idx)
+    for (int idx = 0; idx != d_nStates; ++idx)
       d_prop2[adx][idx] = 1 - accumulate(begin(d_prop1[adx]), end(d_prop1[adx]) - idx - 1, 0.0);
   }
   
@@ -55,7 +55,7 @@ void Init::setTransitionGamma()
 //   
 //   file << "0, " << d_nRates - 1 << '\n';
 //   
-//   for (size_t idx = 0; idx < d_nStates; ++idx)
+//   for (int idx = 0; idx < d_nStates; ++idx)
 //     file << d_prop1[0][idx] << ", " << d_prop1[d_nRates - 1][idx] << '\n';
 //   
 //   throw;
